+++ concur/src-dbcache/dbcache/conf/CacheRule.java
@@ -15,7 +15,7 @@ public class CacheRule {
 	 * @return
 	 */
 	public static Integer getIndexIdKey(String indexName, Object indexValue) {
-		return Integer.valueOf((17 + indexName.hashCode()) * 37 + indexValue.hashCode());
+		return (17 + indexName.hashCode()) * 37 + indexValue.hashCode();
 	}
 
 }
